Added some structure and content
Added first content and a folder structure.
The TOC exists in the _sidebar.md files.I am not completely happy with
this structure and I would prefer a real collapsible sidebar instead of
this clutch where I copy content between the different _sidebar.md files.
This is not a DRY solution at all!