Made the whole repository into a docsify site
I moved the course contents to a directory called 'cursus' and made the
whole repository accessible through Github pages. To make this work both
locally and on the server I had to make the whole repository into a
docsify site and move the index.html up to the root level of the repo.
This also meant changes had to be made to the different _sidebar.md’s.

All this in the hope that I will be able to  create a number of
reveal.js slide decks that will also accessible through Github pages.