Created a template for reveal.js slideshows
Created a template that can be used to build reveal.js markdown
slideshows for this course.

To create a slideshow, copy the template directory and give it the name
of the chapter that it belongs to.

TODO: Set styling and check other configuration parameters of
reveal.js for beautiful, consistent slideshows.