Use npm to manage external dependencies
We are using npm to manage our external dependencies.
In order to be
able to work offline, we link to the external libraries in our
node_modules directory.
Because these links have to work also in our Github repository (through
Github pages), we have added the node_modules directory to the
repository, even though this is not a best practice generally.