Added a README.md as a docsify index of slides
Added a README.md file as a docsify page in the slides directory.
In this way we can create an index of available slides from within the
docsify site for easier navigation.

To make these links work we need to add the :ignore attribute in the
markdown link’s title as per
https://github.com/docsifyjs/docsify/issues/203