Moved the Github pages to a custom domain
I had trouble with the starting point of the URL’s that were not the
same on localhost (/)  and Github pages (webontwikkeling-cursus/).
The easiest way to solve this was setting up a custom domain for the
Github pages (webo.7wn.be), then the URL’s will start from `/` on the
remote site as well.

Since this was already the plan, this was kind of a no-brainer.